pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name            = "focusguard"
version         = "5.1.0"
description     = "Local distraction detector and focus enforcement engine"
readme          = "README.md"
license         = { text = "MIT" }
requires-python = ">=3.9"
keywords        = ["productivity", "focus", "pomodoro", "distraction", "ai", "ocr"]
classifiers     = [
    "Development Status :: 4 - Beta",
    "Environment :: MacOS X",
    "Environment :: Win32 (MS Windows)",
    "Environment :: X11 Applications",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Scheduling",
    "Topic :: Utilities",
]

dependencies = [
    "mss>=9.0.1",
    "Pillow>=10.0.0",
    "opencv-python>=4.9.0",
    "numpy>=1.24.0",
    "easyocr>=1.7.1",
    "pyautogui>=0.9.54",
    "screen-brightness-control>=0.23.0",
    "customtkinter>=5.2.2",
    "requests>=2.31.0",
    "platformdirs>=4.0.0",
]

[project.optional-dependencies]
keyboard = ["pynput>=1.7.6"]          # keyboard idle detection
ambient  = ["sounddevice>=0.4.6"]       # ambient rain/cafe/noise sounds
ollama   = []                          # ollama is a separate binary, not a pip package
dev      = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.urls]
Homepage      = "https://github.com/your-username/focusguard"
Documentation = "https://github.com/your-username/focusguard#readme"
"Bug Tracker" = "https://github.com/your-username/focusguard/issues"
Changelog     = "https://github.com/your-username/focusguard/releases"

[project.scripts]
focusguard     = "focusguard.__main__:main"
focusguard-cli = "focusguard.__main__:main_cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["focusguard*"]

[tool.setuptools.package-data]
focusguard = ["locales/*.py", "assets/*", "logs/.gitkeep"]

[tool.ruff]
line-length    = 100
target-version = "py39"

[tool.ruff.lint]
select  = ["E", "F", "W", "I", "UP", "B"]
ignore  = ["E501", "B008"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.mypy]
python_version         = "3.9"
ignore_missing_imports = true
warn_unused_ignores    = true

[tool.pytest.ini_options]
testpaths    = ["tests"]
addopts      = "-v --tb=short"
filterwarnings = ["ignore::DeprecationWarning"]
